Moved Account components & store into a separate directory
