+++ metaspace/webapp/src/modules/Account/components/CreateAccountDialog.spec.ts
@@ -4,13 +4,13 @@ import ElementUI from 'element-ui';
 import Vuex from 'vuex';
 import Vue from 'vue';
 import CreateAccountDialog from './CreateAccountDialog.vue';
-import accountModule from '../../store/accountModule';
-import router from '../../router';
-import registerMockComponent from '../../../tests/utils/registerMockComponent';
-import { restoreConsole, suppressConsoleWarn } from '../../../tests/utils/suppressConsole';
+import account from '../store/account';
+import router from '../../../router';
+import registerMockComponent from '../../../../tests/utils/registerMockComponent';
+import { restoreConsole, suppressConsoleWarn } from '../../../../tests/utils/suppressConsole';
 
-jest.mock('../../api/auth');
-import * as _mockAuthApi from '../../api/auth';
+jest.mock('../../../api/auth');
+import * as _mockAuthApi from '../../../api/auth';
 const mockAuthApi = _mockAuthApi as jest.Mocked<typeof _mockAuthApi>;
 
 Vue.use(ElementUI);
@@ -39,7 +39,7 @@ describe('CreateAccountDialog', () => {
 
   const store = new Vuex.Store({
     modules: {
-      account: accountModule,
+      account: account,
     },
   });
