+++ metaspace/webapp/src/router.ts
@@ -2,8 +2,8 @@ import Vue from 'vue';
 import VueRouter from 'vue-router';
 import AboutPage from './components/AboutPage.vue';
 import DatasetsPage from './components/DatasetsPage.vue';
-import DialogPage from './components/Account/DialogPage';
-import ResetPasswordPage from './components/Account/ResetPasswordPage.vue';
+import DialogPage from './modules/Account/components/DialogPage';
+import ResetPasswordPage from './modules/Account/components/ResetPasswordPage.vue';
 import * as config from './clientConfig.json';
 
 Vue.use(VueRouter);
